Make the CATCH2 utils available for all tests
